MediAssist-AI-main/backend/requirements.txt
fastapi==0.115.7
uvicorn==0.35.0
python-multipart==0.0.6
pydantic==1.10.7
sqlalchemy==2.0.43
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
openai==1.30.0
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system=="Windows"
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
numpy==1.24.3
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
protobuf==4.25.8
pyarrow==21.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.43
starlette==0.47.2
streamlit==1.47.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
gtts==2.5.1
